feat(all): add some new modules
Promise
otel
template

+++ yggdrasil-bean/src/main/java/io/github/s3s3l/yggdrasil/bean/Sortable.java
@@ -1,5 +1,5 @@
 package io.github.s3s3l.yggdrasil.bean;
 
 public interface Sortable {
-    int getPriority();
+    long getPriority();
 }
