+++ yggdrasil-boot/core/src/main/java/io/github/s3s3l/yggdrasil/boot/processor/AnnotationProcessor.java
@@ -7,12 +7,12 @@ import io.github.s3s3l.yggdrasil.boot.bean.meta.AnnotationMeta;
 
 public interface AnnotationProcessor<T extends Annotation> extends Sortable {
 
-    public static final int DEFAULT_PRIORITY = 0;
+    public static final long DEFAULT_PRIORITY = 0;
     
     AnnotationMeta process(AnnotationMeta pre, T annotation);
 
     @Override
-    default int getPriority() {
+    default long getPriority() {
         return DEFAULT_PRIORITY;
     }
