+++ yggdrasil-boot/spring-adapter/src/main/java/io/github/s3s3l/yggdrasil/boot/spring/adapter/processor/QualifierAnnotationProcessor.java
@@ -16,7 +16,7 @@ public class QualifierAnnotationProcessor extends SpringAnnotationProcessor<Qual
     }
 
     @Override
-    public int getPriority() {
+    public long getPriority() {
         return SpringAnnotationProcessor.PRIORITY + 1;
     }
