+++ yggdrasil-boot/spring-adapter/src/main/java/io/github/s3s3l/yggdrasil/boot/spring/adapter/processor/SpringAnnotationProcessor.java
@@ -5,10 +5,10 @@ import java.lang.annotation.Annotation;
 import io.github.s3s3l.yggdrasil.boot.processor.AnnotationProcessor;
 
 public abstract class SpringAnnotationProcessor<T extends Annotation> implements AnnotationProcessor<T> {
-    public static final int PRIORITY = AnnotationProcessor.DEFAULT_PRIORITY - 1000;
+    public static final long PRIORITY = AnnotationProcessor.DEFAULT_PRIORITY - 1000;
 
     @Override
-    public int getPriority() {
+    public long getPriority() {
         return PRIORITY;
     }
